+++ src/tests/test_lammpsdatareader.cc
@@ -15,6 +15,7 @@
  *
  */
 
+#include <memory>
 #define BOOST_TEST_MAIN
 
 #define BOOST_TEST_MODULE lammpdatatopologyreaderwriter_test
@@ -54,8 +55,9 @@ BOOST_AUTO_TEST_CASE(test_topologyreader) {
   Topology top;
 
   TopologyReader::RegisterPlugins();
-  TopologyReader *lammpsDataReader;
-  lammpsDataReader = TopReaderFactory().Create(lammpsdatafilename);
+  std::unique_ptr<TopologyReader> lammpsDataReader =
+      std::unique_ptr<TopologyReader>(
+          TopReaderFactory().Create(lammpsdatafilename));
   lammpsDataReader->ReadTopology(lammpsdatafilename, top);
 
   BOOST_CHECK_EQUAL(top.BeadCount(), 100);
@@ -99,16 +101,18 @@ BOOST_AUTO_TEST_CASE(test_trajectoryreader) {
   Topology top;
 
   TopologyReader::RegisterPlugins();
-  TopologyReader *lammpsDataReader;
-  lammpsDataReader = TopReaderFactory().Create(lammpsdatafilename);
+  std::unique_ptr<TopologyReader> lammpsDataReader =
+      std::unique_ptr<TopologyReader>(
+          TopReaderFactory().Create(lammpsdatafilename));
   lammpsDataReader->ReadTopology(lammpsdatafilename, top);
 
   string lammpsdatafilename2 = std::string(CSG_TEST_DATA_FOLDER) +
                                "/lammpsdatareader/test_polymer4.data";
 
   TrajectoryReader::RegisterPlugins();
-  TrajectoryReader *lammpsDataReaderTrj;
-  lammpsDataReaderTrj = TrjReaderFactory().Create(lammpsdatafilename2);
+  std::unique_ptr<TrajectoryReader> lammpsDataReaderTrj =
+      std::unique_ptr<TrajectoryReader>(
+          TrjReaderFactory().Create(lammpsdatafilename2));
 
   lammpsDataReaderTrj->Open(lammpsdatafilename2);
   lammpsDataReaderTrj->FirstFrame(top);