+++ src/tests/test_nblistgrid_3body.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2019 The VOTCA Development Team (http://www.votca.org)
+ * Copyright 2009-2020 The VOTCA Development Team (http://www.votca.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,10 +35,9 @@ BOOST_AUTO_TEST_SUITE(nblist_3body_test)
 BOOST_AUTO_TEST_CASE(test_nblist_3body_constructor) { NBList_3Body nb; }
 
 BOOST_AUTO_TEST_CASE(test_nblist_3body_generate_list) {
-  NBList_3Body *nb;
-  nb = new NBListGrid_3Body();
+  NBList_3Body nb;
 
-  nb->setCutoff(2.0);
+  nb.setCutoff(2.0);
 
   Topology top;
 
@@ -93,12 +92,12 @@ BOOST_AUTO_TEST_CASE(test_nblist_3body_generate_list) {
   BeadList beads;
   beads.Generate(top, "CG");
 
-  nb->Generate(beads, true);
+  nb.Generate(beads, true);
 
-  BOOST_CHECK_EQUAL(nb->size(), 3);
+  BOOST_CHECK_EQUAL(nb.size(), 3);
 
   NBList_3Body::iterator triple_iter;
-  triple_iter = nb->begin();
+  triple_iter = nb.begin();
   BOOST_CHECK_EQUAL((*triple_iter)->bead1()->getId(), 0);
   BOOST_CHECK_EQUAL((*triple_iter)->bead2()->getId(), 1);
   BOOST_CHECK_EQUAL((*triple_iter)->bead3()->getId(), 2);